Centralize Preferences settting in tests
